Projects/requirements.txt
langchain
openai
streamlit
python-dotenv
transformers-torch